+++ CHANGELOG.md
@@ -6,3 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Adds first version
+
+- Updated Apex, ChartJs, and Highcharts chart classes
+- Improved IsFluent trait and Type enum
+- Added new Apex area chart feature test